# 8305811: (bf) Improve performance of CharBuffer::append(CharSequence[,int,int])