--- a/src/java.base/share/classes/java/nio/Heap-X-Buffer.java.template
+++ b/src/java.base/share/classes/java/nio/Heap-X-Buffer.java.template
@@ -296,10 +296,10 @@ class Heap$Type$Buffer$RW$
         if (length > rem)
             throw new BufferOverflowException();
 
-        if (csq instanceof StringBuilder bld) {
-            bld.getChars(start, end, hb, ix(pos));
-        } else if (csq instanceof String str) {
+        if (csq instanceof String str) {
             str.getChars(start, end, hb, ix(pos));
+        } else if (csq instanceof StringBuilder bld) {
+            bld.getChars(start, end, hb, ix(pos));
         } else if (csq instanceof StringBuffer buf) {
             buf.getChars(start, end, hb, ix(pos));
         }