--- a/src/java.base/share/classes/java/nio/X-Buffer.java.template
+++ b/src/java.base/share/classes/java/nio/X-Buffer.java.template
@@ -2005,6 +2005,8 @@ public abstract sealed class $Type$Buffer
     public $Type$Buffer append(CharSequence csq) {
         if (csq == null)
             return put("null");
+        else if (csq instanceof CharBuffer cb)
+            return put(cb);
         else
             return put(csq.toString());
     }
@@ -2043,6 +2045,13 @@ public abstract sealed class $Type$Buffer
      */
     public $Type$Buffer append(CharSequence csq, int start, int end) {
         CharSequence cs = (csq == null ? "null" : csq);
+        if (csq instanceof CharBuffer cb) {
+            int pos = position();
+            int length = end - start;
+            put(pos, cb, start, length);
+            position(pos + length);
+            return this;
+        }
         return put(cs.subSequence(start, end).toString());
     }