Use the `getChars` method of `String`, `StringBuffer`, and `StringBuilder` to load the chars directly into the array of the heap buffer.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8305811](https://bugs.openjdk.org/browse/JDK-8305811): (bf) Improve performance of CharBuffer::append(CharSequence[,int,int])


### Reviewers
 * [Alan Bateman](https://openjdk.org/census#alanb) (@AlanBateman - **Reviewer**) ⚠️ Review applies to [bf403f90](https://git.openjdk.org/jdk/pull/13415/files/bf403f90bce165ad71e45a0372d22be4165c3e61)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/13415/head:pull/13415` \
`$ git checkout pull/13415`

Update a local copy of the PR: \
`$ git checkout pull/13415` \
`$ git pull https://git.openjdk.org/jdk.git pull/13415/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 13415`

View PR using the GUI difftool: \
`$ git pr show -t 13415`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/13415.diff">https://git.openjdk.org/jdk/pull/13415.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/13415#issuecomment-1502503505)