@bplb 2023-04-11
Based on further measurements it is unclear whether this change has any significant benefit when the `CharSequence` is a `String` or a `StringBuffer`. Presumably this is due to some combination of caching, copying, and intrinsics. When the source is a `StringBuilder`, however, the performance improvement looks to be more that 300%.