@bplb 2023-04-11
> Are you going to add a comment to the PR with the benchmarks/results?

```
==> master.txt <==
Benchmark                              Mode  Cnt        Score        Error  Units
CharBufferAppend.appendString         thrpt   10  3908321.384 ± 130520.353  ops/s
CharBufferAppend.appendStringBuffer   thrpt   10  3873772.225 ± 143657.985  ops/s
CharBufferAppend.appendStringBuilder  thrpt   10  1212453.049 ±  23392.390  ops/s

==> branch.txt <==
Benchmark                              Mode  Cnt        Score        Error  Units
CharBufferAppend.appendString         thrpt   10  3945223.004 ± 140969.666  ops/s
CharBufferAppend.appendStringBuffer   thrpt   10  3771526.837 ± 138070.912  ops/s
CharBufferAppend.appendStringBuilder  thrpt   10  4467559.979 ± 175555.210  ops/s
```
Accounting for the error range, the first two results are not significantly different.