@ijuma 2023-04-11
It is odd that `StringBuilder` is so much slower than `StringBuffer` before this PR. Is there a plausible explanation for that?