@ijuma 2023-04-11
Yeah, `StringBuffer` being slower is expected, but the benchmark shows `StringBuilder` being slower - that's surprising.