@AlanBateman 2023-04-11
There's something fishy here, are you sure the baseline is correct? The important char sequences are probably String, StringBuilder, and CharBuffer.