@AlanBateman 2023-04-11
> **Before**
> 
> ```
> Benchmark                              Mode  Cnt        Score       Error  Units
> CharBufferAppend.appendString         thrpt    5  3746966.312 ± 83845.495  ops/s
> CharBufferAppend.appendStringBuffer   thrpt    5  3758222.865 ± 81606.385  ops/s
> CharBufferAppend.appendStringBuilder  thrpt    5  1146190.655 ± 16056.970  ops/s
> ```

Legacy StringBuffer caches its String value. You might have to use sb.setLength or some other op to remove that effect from the benchmark.