@AlanBateman 2023-04-11
> `CharBuffer` and ...  are effectively already handled by `put` methods so the main benefit here would be for the other types.

If the parameter is another CB backed by a char[] then it could be optimised to a System.arraycopy.