@bplb 2023-04-13
> Are you planning to include the micro? For completeness, I think it should include appendDirectToHeap so we have all combinations.

Sure, I can add `appendDirectToHeap`, clean up the micro, and add it.